Refactor customer and station management in backend
- Updated SQL schema for customer and station tables.
- Renamed and modified DTOs for customer creation.
- Refactored exception handling to use consistent package naming.
- Improved CustomerService and StaffService with async methods.
- Added CustomerController to manage customer-related endpoints.
- Updated tests for customer and station functionalities.
- Removed unused service classes and methods.

+++ Backend/src/main/java/org/coffeeshop/stations/repositories/StationRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StationRepository extends JpaRepository<Station, Integer> {}
+public interface StationRepository extends JpaRepository<Station, Long> {}
